auth middleware added to categories routes

+++ backend/controllers/transaction.category.controller.js
@@ -1,12 +1,15 @@
 import asyncHandler from 'express-async-handler';
 import Category from '../models/transaction.category.model.js';
+import User from '../models/user.model.js';
 import AppError from '../utils/appError.js';
 
 //@desc    Get all categories
 //@route   GET /api/v1/categories
 //@access  Private
 export const getAllCategories = asyncHandler(async (req, res) => {
-	const categories = await Category.find().sort({ createdAt: -1 });
+	const categories = await Category.find({ user: req.user._id }).sort({
+		createdAt: -1,
+	});
 	res.status(200).json({ success: true, data: categories });
 });
 
@@ -26,6 +29,7 @@ export const addCategory = asyncHandler(async (req, res) => {
 
 	const newCategory = await Category.create({
 		name,
+		user: req.user._id,
 		type,
 		color,
 	});
@@ -44,6 +48,18 @@ export const updateCategory = asyncHandler(async (req, res) => {
 	if (!existingCategory) {
 		throw new AppError('Category not found', 404);
 	}
+
+	const user = await User.findById(req.user._id);
+	// Find whether user is logged in user
+	if (!user) {
+		throw new AppError('User not found. Please Sign-in', 401);
+	}
+
+	// Check if the logged in user is the owner of the category
+	if (existingCategory.user.toString() !== user._id.toString()) {
+		throw new AppError('User not authorized to update this category.', 403);
+	}
+
 	const { name, type, color } = req.body;
 	const updatedCategory = await Category.findByIdAndUpdate(
 		req.params.id,
@@ -68,6 +84,17 @@ export const deleteCategory = asyncHandler(async (req, res) => {
 	const existingCategory = await Category.findById(req.params.id);
 	if (!existingCategory) {
 		throw new AppError('Category not found', 404);
+	}
+
+	const user = await User.findById(req.user._id);
+	// Find whether user is logged in user
+	if (!user) {
+		throw new AppError('User not found. Please Sign-in', 401);
+	}
+
+	// Check if the logged in user is the owner of the category
+	if (existingCategory.user.toString() !== user._id.toString()) {
+		throw new AppError('User not authorized to update this category', 403);
 	}
 	await Category.findByIdAndDelete(req.params.id);
 	res.status(200).json({
